split App.tsx & combine ImplWrapper props
