+++ packages/runtime/src/App.tsx
@@ -1,435 +1,32 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import {
-  Application,
-  createApplication,
-  RuntimeApplication,
-} from '@meta-ui/core';
-import { merge } from 'lodash';
-import { Registry, TraitResult } from './registry';
-import { StateManager } from './store';
-import { ApiService } from './api-service';
-import { ContainerPropertySchema } from './traits/core/slot';
-import { Static } from '@sinclair/typebox';
-import { watch } from '@vue-reactivity/watch';
-import copy from 'copy-to-clipboard';
-import { GlobalHandlerMap } from './handler';
+import React, { useMemo } from 'react';
+import { createApplication } from '@meta-ui/core';
 import { initStateAndMethod } from './utils/initStateAndMethod';
+import { ImplWrapper } from './modules/ImplWrapper';
+import { resolveAppComponents } from './modules/resolveAppComponents';
+import { AppProps, MetaUIModules } from './types/RuntimeSchema';
+import { DebugEvent, DebugStore } from './modules/DebugComponents';
 
-type ArrayElement<ArrayType extends readonly unknown[]> =
-  ArrayType extends readonly (infer ElementType)[] ? ElementType : never;
-
-type ApplicationComponent = RuntimeApplication['spec']['components'][0];
-type ApplicationTrait = ArrayElement<ApplicationComponent['traits']>;
-
-type ComponentWrapperType = React.FC<{ id: string }>;
-
-export const ImplWrapper = React.forwardRef<
-  HTMLDivElement,
-  {
-    component: ApplicationComponent;
-    slotsMap: SlotsMap | undefined;
-    targetSlot: { id: string; slot: string } | null;
-    app?: RuntimeApplication;
-    registry: Registry;
-    stateManager: StateManager;
-    globalHandlerMap: GlobalHandlerMap;
-    apiService: ApiService;
-    componentWrapper?: ComponentWrapperType;
-    // [key: string]: any;
-  }
->(
-  (
-    {
-      component: c,
-      slotsMap,
-      targetSlot,
-      app,
-      registry,
-      stateManager,
-      globalHandlerMap,
-      apiService,
-      children,
-      componentWrapper: ComponentWrapper,
-      ...props
-    },
-    ref
-  ) => {
-    const Impl = registry.getComponent(
-      c.parsedType.version,
-      c.parsedType.name
-    ).impl;
-
-    if (!globalHandlerMap.has(c.id)) {
-      globalHandlerMap.set(c.id, {});
-    }
-
-    let handlerMap = globalHandlerMap.get(c.id)!;
-    useEffect(() => {
-      const handler = (s: {
-        componentId: string;
-        name: string;
-        parameters?: any;
-      }) => {
-        if (s.componentId !== c.id) {
-          return;
-        }
-        if (!handlerMap[s.name]) {
-          // maybe log?
-          return;
-        }
-        handlerMap[s.name](s.parameters);
-      };
-      apiService.on('uiMethod', handler);
-      return () => {
-        apiService.off('uiMethod', handler);
-        globalHandlerMap.delete(c.id);
-      };
-    }, []);
-
-    const mergeState = useCallback(
-      (partial: any) => {
-        stateManager.store[c.id] = { ...stateManager.store[c.id], ...partial };
-      },
-      [c.id]
-    );
-    const subscribeMethods = useCallback((map: any) => {
-      handlerMap = { ...handlerMap, ...map };
-      globalHandlerMap.set(c.id, handlerMap);
-    }, []);
-
-    // result returned from traits
-    const [traitResults, setTraitResults] = useState<TraitResult[]>(() => {
-      return c.traits.map(trait =>
-        excecuteTrait(trait, stateManager.deepEval(trait.properties).result)
-      );
-    });
-
-    function excecuteTrait(
-      trait: ApplicationTrait,
-      traitProperty: ApplicationTrait['properties']
-    ) {
-      const tImpl = registry.getTrait(
-        trait.parsedType.version,
-        trait.parsedType.name
-      ).impl;
-      return tImpl({
-        ...traitProperty,
-        componentId: c.id,
-        mergeState,
-        subscribeMethods,
-        stateManager,
-        apiService,
-      });
-    }
-
-    // eval traits' properties then excecute traits
-    useEffect(() => {
-      const stops: ReturnType<typeof watch>[] = [];
-      const properties: Array<ApplicationTrait['properties']> = [];
-      c.traits.forEach((t, i) => {
-        const { result, stop } = stateManager.deepEval(
-          t.properties,
-          ({ result: property }: any) => {
-            const traitResult = excecuteTrait(t, property);
-            setTraitResults(oldResults => {
-              // assume traits number and order will not change
-              const newResults = [...oldResults];
-              newResults[i] = traitResult;
-              return newResults;
-            });
-            stops.push(stop);
-          }
-        );
-        properties.push(result);
-      });
-      // although traitResults has initialized in useState, it must be set here again
-      // because mergeState will be called during the first render of component, and state will change
-      setTraitResults(
-        c.traits.map((trait, i) => excecuteTrait(trait, properties[i]))
-      );
-      return () => stops.forEach(s => s());
-    }, [c.traits]);
-
-    // reduce traitResults
-    const propsFromTraits: TraitResult['props'] = useMemo(() => {
-      return Array.from(traitResults.values()).reduce(
-        (prevProps, result: TraitResult) => {
-          if (!result.props) {
-            return prevProps;
-          }
-
-          let effects = prevProps?.effects || [];
-          if (result.props?.effects) {
-            effects = effects?.concat(result.props?.effects);
-          }
-          return {
-            ...prevProps,
-            ...result.props,
-            effects,
-          };
-        },
-        {} as TraitResult['props']
-      );
-    }, [traitResults]);
-
-    // component properties
-    const [evaledComponentProperties, setEvaledComponentProperties] = useState(
-      () => {
-        return merge(
-          stateManager.deepEval(c.properties).result,
-          propsFromTraits
-        );
-      }
-    );
-
-    // eval component properties
-    useEffect(() => {
-      const { result, stop } = stateManager.deepEval(
-        c.properties,
-        ({ result: newResult }: any) => {
-          setEvaledComponentProperties({ ...newResult });
-        }
-      );
-      // must keep this line, reason is the same as above
-      setEvaledComponentProperties({ ...result });
-      return stop;
-    }, [c.properties]);
-
-    const mergedProps = { ...evaledComponentProperties, ...propsFromTraits };
-
-    const C = (
-      <Impl
-        key={c.id}
-        component={c}
-        {...mergedProps}
-        registry={registry}
-        stateManager={stateManager}
-        globalHandlerMap={globalHandlerMap}
-        apiService={apiService}
-        mergeState={mergeState}
-        subscribeMethods={subscribeMethods}
-        slotsMap={slotsMap}
-        app={app}
-      />
-    );
-
-    if (targetSlot && app) {
-      const targetC = app.spec.components.find(c => c.id === targetSlot.id);
-      if (targetC?.parsedType.name === 'grid_layout') {
-        return (
-          <div key={c.id} data-meta-ui-id={c.id} ref={ref} {...props}>
-            {C}
-            {children}
-          </div>
-        );
-      }
-    }
-
-    if (ComponentWrapper) {
-      return (
-        <React.Fragment key={c.id}>
-          <ComponentWrapper id={c.id}>
-            {C}
-            {children}
-          </ComponentWrapper>
-        </React.Fragment>
-      );
-    } else {
-      return (
-        <React.Fragment key={c.id}>
-          {C}
-          {children}
-        </React.Fragment>
-      );
-    }
-  }
-);
-
-const DebugStore: React.FC<{ stateManager: StateManager }> = ({
-  stateManager,
-}) => {
-  const [store, setStore] = useState(stateManager.store);
-  useEffect(() => {
-    setStore({ ...stateManager.store });
-    watch(stateManager.store, newValue => {
-      setTimeout(() => {
-        setStore({ ...newValue });
-      }, 0);
-    });
-  }, []);
-
-  return <pre>{JSON.stringify(store, null, 2)}</pre>;
-};
-
-const DebugEvent: React.FC<{ apiService: ApiService }> = ({ apiService }) => {
-  const [events, setEvents] = useState<unknown[]>([]);
-
-  useEffect(() => {
-    const handler = (type: string, event: unknown) => {
-      setEvents(cur => cur.concat({ type, event, t: new Date() }));
-    };
-    apiService.on('*', handler);
-    return () => apiService.off('*', handler);
-  }, []);
-
-  return (
-    <div>
-      <div>
-        <button
-          onClick={() => {
-            copy(JSON.stringify(events));
-          }}
-        >
-          copy test case
-        </button>
-      </div>
-      <div
-        style={{
-          padding: '0.5em',
-          border: '2px solid black',
-          maxHeight: '200px',
-          overflow: 'auto',
-        }}
-      >
-        {events.map((event, idx) => (
-          <pre key={idx}>{JSON.stringify(event)}</pre>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export type SlotComponentMap = Map<string, SlotsMap>;
-export type SlotsMap = Map<
-  string,
-  Array<{
-    component: React.FC;
-    id: string;
-  }>
->;
-
-export function resolveAppComponents(
-  registry: Registry,
-  stateManager: StateManager,
-  globalHandlerMap: GlobalHandlerMap,
-  apiService: ApiService,
-  components: RuntimeApplication['spec']['components'],
-  componentWrapper?: ComponentWrapperType,
-  app?: RuntimeApplication
-): {
-  topLevelComponents: RuntimeApplication['spec']['components'];
-  slotComponentsMap: SlotComponentMap;
-} {
-  const topLevelComponents: RuntimeApplication['spec']['components'] = [];
-  const slotComponentsMap: SlotComponentMap = new Map();
-
-  for (const c of components) {
-    // handle component with slot trait
-    const slotTrait = c.traits.find(t => t.parsedType.name === 'slot');
-    if (slotTrait) {
-      const { id, slot } = (
-        slotTrait.properties as {
-          container: Static<typeof ContainerPropertySchema>;
-        }
-      ).container;
-      if (!slotComponentsMap.has(id)) {
-        slotComponentsMap.set(id, new Map());
-      }
-      if (!slotComponentsMap.get(id)?.has(slot)) {
-        slotComponentsMap.get(id)?.set(slot, []);
-      }
-      const component = React.forwardRef<HTMLDivElement, any>((props, ref) => (
-        <ImplWrapper
-          component={c}
-          slotsMap={slotComponentsMap.get(c.id)}
-          targetSlot={{ id, slot }}
-          registry={registry}
-          stateManager={stateManager}
-          globalHandlerMap={globalHandlerMap}
-          apiService={apiService}
-          app={app}
-          componentWrapper={componentWrapper}
-          {...props}
-          ref={ref}
-        />
-      ));
-      component.displayName = c.parsedType.name;
-      slotComponentsMap.get(id)?.get(slot)?.push({
-        component,
-        id: c.id,
-      });
-    }
-
-    // if the component is neither assigned with slot trait nor route trait, consider it as a top level component
-    !slotTrait && topLevelComponents.push(c);
-  }
-
-  return {
-    topLevelComponents,
-    slotComponentsMap,
+// inject modules to App
+export function genApp(mModules: MetaUIModules) {
+  return (props: Omit<AppProps, 'mModules'>) => {
+    return <App {...props} mModules={mModules} />;
   };
 }
 
-type AppProps = {
-  options: Application;
-  registry: Registry;
-  stateManager: StateManager;
-  globalHandlerMap: GlobalHandlerMap;
-  apiService: ApiService;
-  componentWrapper?: ComponentWrapperType;
-  debugStore?: boolean;
-  debugEvent?: boolean;
-};
-
-export function genApp(
-  registry: Registry,
-  stateManager: StateManager,
-  globalHandlerMap: GlobalHandlerMap,
-  apiService: ApiService
-) {
-  return (
-    props: Omit<
-      AppProps,
-      'registry' | 'stateManager' | 'globalHandlerMap' | 'apiService'
-    >
-  ) => {
-    return (
-      <App
-        {...props}
-        registry={registry}
-        stateManager={stateManager}
-        globalHandlerMap={globalHandlerMap}
-        apiService={apiService}
-      />
-    );
-  };
-}
-
-export const App: React.FC<AppProps> = ({
-  options,
-  registry,
-  stateManager,
-  globalHandlerMap,
-  apiService,
-  componentWrapper,
-  debugStore = true,
-  debugEvent = true,
-}) => {
+export const App: React.FC<AppProps> = props => {
+  const {
+    options,
+    mModules,
+    componentWrapper,
+    debugStore = true,
+    debugEvent = true,
+  } = props;
   const app = createApplication(options);
 
-  initStateAndMethod(registry, stateManager, app.spec.components);
+  initStateAndMethod(mModules.registry, mModules.stateManager, app.spec.components);
 
   const { topLevelComponents, slotComponentsMap } = useMemo(
-    () =>
-      resolveAppComponents(
-        registry,
-        stateManager,
-        globalHandlerMap,
-        apiService,
-        app.spec.components,
-        componentWrapper,
-        app
-      ),
+    () => resolveAppComponents(mModules, app.spec.components, componentWrapper, app),
     [app]
   );
 
@@ -440,10 +37,7 @@ export const App: React.FC<AppProps> = ({
           <ImplWrapper
             key={c.id}
             component={c}
-            registry={registry}
-            stateManager={stateManager}
-            globalHandlerMap={globalHandlerMap}
-            apiService={apiService}
+            mModules={mModules}
             slotsMap={slotComponentsMap.get(c.id)}
             targetSlot={null}
             app={app}
@@ -451,8 +45,8 @@ export const App: React.FC<AppProps> = ({
           />
         );
       })}
-      {debugStore && <DebugStore stateManager={stateManager} />}
-      {debugEvent && <DebugEvent apiService={apiService} />}
+      {debugStore && <DebugStore stateManager={mModules.stateManager} />}
+      {debugEvent && <DebugEvent apiService={mModules.apiService} />}
     </div>
   );
 };
