+++ packages/runtime/src/components/chakra-ui/HStack.tsx
@@ -1,7 +1,7 @@
 import { createComponent } from '@meta-ui/core';
 import { Static, Type } from '@sinclair/typebox';
 import { HStack as BaseHStack } from '@chakra-ui/react';
-import { ComponentImplementation } from '../../registry';
+import { ComponentImplementation } from '../../modules/registry';
 import Slot from '../_internal/Slot';
 import {
   DirectionSchema,
