+++ packages/runtime/src/components/chakra-ui/Table/Table.tsx
@@ -1,17 +1,9 @@
 import { useEffect, useMemo, useState } from 'react';
 import { sortBy } from 'lodash';
-import {
-  Table as BaseTable,
-  Thead,
-  Tr,
-  Th,
-  Tbody,
-  Checkbox,
-  Td,
-} from '@chakra-ui/react';
+import { Table as BaseTable, Thead, Tr, Th, Tbody, Checkbox, Td } from '@chakra-ui/react';
 import { Static } from '@sinclair/typebox';
 import { TablePagination } from './Pagination';
-import { ComponentImplementation } from '../../../registry';
+import { ComponentImplementation } from '../../../modules/registry';
 import {
   ColumnsPropertySchema,
   DataPropertySchema,
@@ -42,19 +34,14 @@ export const TableImpl: ComponentImplementation<{
   columns,
   isMultiSelect,
   mergeState,
-  stateManager,
-  apiService,
+  mModules,
 }) => {
   if (!data) {
     return <div>loading</div>;
   }
   const normalizedData = data;
-  const [selectedItem, setSelectedItem] = useState<
-    Record<string, any> | undefined
-  >();
-  const [selectedItems, setSelectedItems] = useState<
-    Array<Record<string, any>>
-  >([]);
+  const [selectedItem, setSelectedItem] = useState<Record<string, any> | undefined>();
+  const [selectedItems, setSelectedItems] = useState<Array<Record<string, any>>>([]);
   const [currentPage, setCurrentPage] = useState<number>(0);
   const [sortRule, setSortRule] = useState<SortRule | undefined>();
   const pageNumber = Math.ceil(data.length / rowsPerPage);
@@ -90,10 +77,7 @@ export const TableImpl: ComponentImplementation<{
 
   function isItemSelected(target: any) {
     if (isMultiSelect) {
-      return (
-        selectedItems.findIndex(item => item[majorKey] === target[majorKey]) >
-        -1
-      );
+      return selectedItems.findIndex(item => item[majorKey] === target[majorKey]) > -1;
     }
     return selectedItem && selectedItem[majorKey] === target[majorKey];
   }
@@ -114,8 +98,7 @@ export const TableImpl: ComponentImplementation<{
   }
 
   const allCheckbox = useMemo(() => {
-    const isAllChecked =
-      isMultiSelect && selectedItems.length === normalizedData.length;
+    const isAllChecked = isMultiSelect && selectedItems.length === normalizedData.length;
     const isIndeterminate =
       selectedItems.length > 0 && selectedItems.length < normalizedData.length;
     const onChange = (e: any) => {
@@ -199,8 +182,7 @@ export const TableImpl: ComponentImplementation<{
                     item={item}
                     column={column}
                     onClickItem={() => selectItem(item)}
-                    stateManager={stateManager}
-                    apiService={apiService}
+                    mModules={mModules}
                   />
                 ))}
               </Tr>
