+++ packages/runtime/src/index.ts
@@ -1,9 +1,9 @@
-import { initStateManager } from './store';
+import { initStateManager } from './modules/stateStore';
 import { genApp } from './App';
-import { initRegistry } from './registry';
-import { mountUtilMethods } from './util-methods';
-import { initGlobalHandlerMap } from './handler';
-import { initApiService } from './api-service';
+import { initRegistry } from './modules/registry';
+import { initApiService } from './modules/apiService';
+import { mountUtilMethods } from './modules/util-methods';
+import { initGlobalHandlerMap } from './modules/handler';
 
 export function initMetaUI() {
   const registry = initRegistry();
@@ -13,7 +13,7 @@ export function initMetaUI() {
   mountUtilMethods(apiService);
 
   return {
-    App: genApp(registry, stateManager, globalHandlerMap, apiService),
+    App: genApp({ registry, stateManager, globalHandlerMap, apiService }),
     stateManager,
     registry,
     globalHandlerMap,
