+++ packages/runtime/src/utils/initStateAndMethod.ts
@@ -1,7 +1,7 @@
 import { RuntimeApplication } from '@meta-ui/core';
 import { TSchema } from '@sinclair/typebox';
-import { Registry } from 'src/registry';
-import { StateManager } from 'src/store';
+import { Registry } from 'src/modules/registry';
+import { StateManager } from 'src/modules/stateStore';
 import { parseTypeBox } from './parseTypeBox';
 
 export function initStateAndMethod(
@@ -15,16 +15,10 @@ export function initStateAndMethod(
     }
     let state = {};
     c.traits.forEach(t => {
-      const tSpec = registry.getTrait(
-        t.parsedType.version,
-        t.parsedType.name
-      ).spec;
+      const tSpec = registry.getTrait(t.parsedType.version, t.parsedType.name).spec;
       state = { ...state, ...parseTypeBox(tSpec.state as TSchema) };
     });
-    const cSpec = registry.getComponent(
-      c.parsedType.version,
-      c.parsedType.name
-    ).spec;
+    const cSpec = registry.getComponent(c.parsedType.version, c.parsedType.name).spec;
     state = { ...state, ...parseTypeBox(cSpec.state as TSchema) };
     stateManager.store[c.id] = state;
   });
